requirements.txt
nltk==3.9.1
openai-whisper==20240930
pydub==0.25.1
pytest==8.3.5
torch==2.6.0
